backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
google-auth
google-auth-oauthlib
apscheduler
